+++ stock_exit/views.py
@@ -1,16 +1,16 @@
 from django.shortcuts import get_object_or_404
 from .forms import StockExitForm
 from django.views.generic import ListView, CreateView, UpdateView, DeleteView
-from django.urls import reverse
 from .models import StockExit
 from django.http import JsonResponse
 from django.contrib import messages
 
 
 class StockExitListView(ListView):
-    def post(self, request):
-        stock_exit = StockExit.objects.all()
-        response = {"data": [
+    @staticmethod
+    def post(request):
+        stock_exit = StockExit.objects.filter(user=request.user)
+        data = {"data": [
             {
               "DT_RowId": f"row_{exit.id}",
               "Data da Venda": exit.get_link_update(),
@@ -25,43 +25,62 @@ class StockExitListView(ListView):
             for exit in stock_exit
             ]
         }
-        return JsonResponse(response)
+        return JsonResponse(data=data, status=200)
 
 
 class StockExitCreateView(CreateView):
     template_name = 'stock_exit/form.html'
     form_class = StockExitForm
+    success_message = 'Saída cadastrada com sucesso !'
+    error_message = 'Erro ao cadastra saída !'
+
+    def get_initial(self, *args, **kwargs):
+        initial = super(StockExitCreateView, self).get_initial(**kwargs)
+        initial['user'] = self.request.user
+        return initial
 
     def form_valid(self, form):
+        form.instance.user = self.request.user
+        messages.success(self.request, self.success_message)
         return super().form_valid(form)
 
-    def get_success_url(self):
-        message = "Saída cadastrada com sucesso !"
-        messages.add_message(self.request, messages.SUCCESS, message)
-        return reverse('stock_exit:main')
+    def form_invalid(self, form):
+        messages.error(self.request, self.error_message)
+        return super().form_invalid(form)
 
 
 class StockExitUpdateView(UpdateView):
     template_name = 'stock_exit/form.html'
     form_class = StockExitForm
+    success_message = 'Saída atualizar com sucesso !'
+    error_message = 'Erro ao atualizar saída !'
 
-    def get_object(self):
-        id = self.kwargs.get('id')
-        return get_object_or_404(StockExit, id=id)
+    def get_initial(self, *args, **kwargs):
+        initial = super(StockExitUpdateView, self).get_initial(**kwargs)
+        initial['user'] = self.request.user
+        return initial
+
+    def get_object(self, **kwargs):
+        exit_id = self.kwargs.get('id')
+        return get_object_or_404(StockExit, id=exit_id, user=self.request.user)
 
     def form_valid(self, form):
+        messages.success(self.request, self.success_message)
         return super().form_valid(form)
 
-    def get_success_url(self):
-        message = "Saída atualizada com sucesso !"
-        messages.add_message(self.request, messages.SUCCESS, message)
-        return reverse('stock_exit:main')
+    def form_invalid(self, form):
+        messages.error(self.request, self.error_message)
+        return super().form_invalid(form)
 
 
 class StockExitDeleteView(DeleteView):
-    def post(self, request):
-        id = request.POST.get('id')
-        exit = StockExit.objects.get(id=id)
-        exit.delete()
-        response = {'message': 'Saída deletada com sucesso !'}
-        return JsonResponse(response)
+    @staticmethod
+    def post(request, **kwargs):
+        exit_id = request.POST.get('id')
+        exit = StockExit.objects.filter(id=exit_id, user=request.user).first()
+        if exit:
+            exit.delete()
+            data = {'type': 'success', 'message': 'Saída deletada com sucesso !'}
+        else:
+            data = {'type': 'danger', 'message': 'Saída inexistente !'}
+        return JsonResponse(data=data, status=200)
